breaking: New parsers, new definitions, new structure, new types, global imu configs and more

+++ src/parsing.rs
@@ -2,7 +2,7 @@ pub mod upper_sensor;
 pub mod lower_sensor;
 
 
-pub fn fixed_dec<T: Into<f64> + Clone>(shift: f64, v: &T) -> f64 {
+pub fn fixed_dec<T: Into<f32> + Clone>(shift: f32, v: &T) -> f32 {
     (*v).clone().into() / shift
 }
